% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/54_metabolite_origin_analysis.R
\name{analyze_metabolite_origins}
\alias{analyze_metabolite_origins}
\title{Filter a dataset to include only metabolites of known origin}
\usage{
analyze_metabolite_origins(object)
}
\arguments{
\item{object}{A data.frame or mass_dataset object containing metabolite annotation data}
}
\value{
The filtered object with the same class as the input
}
\description{
This function filters a mass dataset to keep only metabolites with known
origin information, removing entries where the 'from_human' field is NA.
}
\author{
Xiaotao Shen \email{xiaotao.shen@outlook.com}
}
