% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotate_result}
\alias{annotate_result}
\title{A metIdentifyClass object.}
\format{
A metIdentifyClass object.
}
\usage{
annotate_result
}
\description{
A metIdentifyClass object.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
\keyword{datasets}
