% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_mona_database.R
\name{construct_mona_database}
\alias{construct_mona_database}
\title{Construct public MS2 database from MoNA with msp format.}
\usage{
construct_mona_database(
  file,
  only.remain.ms2 = TRUE,
  path = ".",
  version = "0.0.1",
  source = "MoNA",
  link = "https://mona.fiehnlab.ucdavis.edu/",
  creater = "Xiaotao Shen",
  email = "shenxt1990@163.com",
  rt = FALSE,
  threads = 5
)
}
\arguments{
\item{file}{The file name of MassBank or MoNA database (mgf format).}

\item{only.remain.ms2}{Only remain the metabolites with MS2 spectra?}

\item{path}{Work directory.}

\item{version}{The version of you database. Default is 0.0.1.}

\item{source}{The source of your database.}

\item{link}{Website link of the source.}

\item{creater}{Creater name. For example, Xiaotao Shen.}

\item{email}{email address.}

\item{rt}{Do the metabolites have RT information or not?. If not, set it as FALSE.}

\item{threads}{The number of threads}
}
\value{
A databaseClass object.
}
\description{
Construct MS2 spectra database according to mzXML data and compound information table (csv format).
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
