% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo1}
\alias{demo1}
\title{Example of input dataset.}
\format{A data frame with 20 rows and 7 variables:
\describe{
  \item{query_m.z}{Mass of compounds.}
  \item{name}{Names of putative IDs.}
  \item{formula}{Formulas of putative IDs.}
  \item{exact_m.z}{Exact mass of putative IDs.}
  \item{pubchem_cid}{PubChem IDs of putative IDs.}
  \item{kegg_id}{KEGG IDs of putative IDs.}
  ...
}}
\usage{
demo1
}
\description{
A dataset which can be used as input dataset and its row names do not match the default row names.
}
\keyword{datasets}
