% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo2}
\alias{demo2}
\title{Example of input dataset, in which colnames does not meet requirement.}
\format{A data frame with 3592 rows and 6 variables:
\describe{
  \item{Query.Mass}{Mass of compounds.}
  \item{Name}{Names of putative IDs.}
  \item{Formula}{Formulas of putative IDs.}
  \item{Exact.Mass}{Exact mass of putative IDs.}
  \item{PubChem.CID}{PubChem IDs of putative IDs.}
  \item{KEGG.ID}{KEGG IDs of putative IDs.}
  ...
}}
\usage{
demo2
}
\description{
A dataset which can be used as input dataset and its row names do not match the default row names.
}
\keyword{datasets}
