% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4metIdentifyClass.R
\name{filter_adducts}
\alias{filter_adducts}
\title{Filter adducts.}
\usage{
filter_adducts(object, remove_adduct = NULL)
}
\arguments{
\item{object}{metIdentifyClass.}

\item{remove_adduct}{What adduct you want to remove from annotation result. Like '(M-H)-'. All the adduct list can be found here:
data("hilic.pos", package = 'metID'), data("hilic.neg", package = 'metID'), 
data("rp.pos", package = 'metID'), data("rp.neg", package = 'metID').}
}
\value{
A MS2 spectrum.
}
\description{
Filter adducts.
\lifecycle{maturing}
}
\examples{
data("annotate_result", package = "metID")
peak_name <- which_has_identification(annotate_result)
get_ms2_spectrum_from_object(annotate_result, peak_name[1,1])
}
\seealso{
The example and demo data of this function can be found
https://jaspershen.github.io/metID/articles/metID.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
