% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentifyClass-methods.R
\name{filter_identification}
\alias{filter_identification}
\title{Filter identifications according to m/z error, RT error, MS similarity and total score}
\usage{
filter_identification(
  object,
  ms1.match.ppm = 25,
  rt.match.tol = 30,
  ms2.match.tol = 0.5,
  total.score.tol = 0.5
)
}
\arguments{
\item{object}{A metIdentifyClass object.}

\item{ms1.match.ppm}{MS1 match ppm.}

\item{rt.match.tol}{RT match tolerance.}

\item{ms2.match.tol}{MS2 match (MS2 similarity) tolerance.}

\item{total.score.tol}{Total score tolerance.}
}
\value{
A new metIdentifyClass.
}
\description{
Filter identifications according to m/z error, RT error, MS similarity and total score.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
