% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentifyClass-methods.R
\name{get_iden_info}
\alias{get_iden_info}
\title{Get identification information from a metIdentifyClass object}
\usage{
get_iden_info(object, which.peak, database)
}
\arguments{
\item{object}{A metIdentifyClass object.}

\item{which.peak}{A peak name or "all". "all" means all peaks with identifications will be output.}

\item{database}{Database used.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification information from a metIdentifyClass object.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
