% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_identification_table.R
\name{get_identification_table}
\alias{get_identification_table}
\title{Get identification table from a metIdentifyClass object}
\usage{
get_identification_table(..., candidate.num = 3, type = c("old", "new"))
}
\arguments{
\item{...}{One or multiple metIdentifyClass objects.}

\item{candidate.num}{The number of candidates.}

\item{type}{The type of identification table.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification table from a metIdentifyClass object.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
