% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_identification_table_all.R
\name{get_identification_table_all}
\alias{get_identification_table_all}
\title{get_identification_table_all}
\usage{
get_identification_table_all(
  ...,
  candidate.num = 1,
  level_condition = c(`mz&rt&ms2` = 1, `mz&rt` = 2, `mz&ms2` = 2, mz = 3)
)
}
\arguments{
\item{...}{One or multiple results from `identify_metabolite_all()`.}

\item{candidate.num}{candidate.num}

\item{level_condition}{Condition for level assign.}
}
\value{
A identification table (data.frame).
}
\description{
Extract the identifications from multiple results of `identify_metabolite_all()`.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
