% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNetworks.R
\name{get_kegg_network}
\alias{get_kegg_network}
\title{Build network between identifications based on kegg network database.}
\usage{
get_kegg_network(kegg_id)
}
\arguments{
\item{kegg_id}{a vector of strings indicating KEGG ID of putative ID.}
}
\value{
a binary matrix of network of KEGG IDs.
}
\description{
Build network between identifications based on kegg network database.
}
