% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNetworks.R
\name{get_tani_network}
\alias{get_tani_network}
\title{Build network between identifications based on tanimoto score.}
\usage{
get_tani_network(pubchem_cid)
}
\arguments{
\item{pubchem_cid}{a vector of strings indicating PubChem CID of putative ID.}
}
\value{
a binary matrix of network of tanimoto scores.
}
\description{
Build network between identifications based on tanimoto score.
}
