% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzIdentify.R
\name{getIdentificationTable2}
\alias{getIdentificationTable2}
\title{getIdentificationTable2}
\usage{
getIdentificationTable2(object, candidate.num = 3, type = c("old", "new"))
}
\arguments{
\item{object}{mzIdentifyClass object.}

\item{candidate.num}{The number of candidates.}

\item{type}{The type of identification table.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification table from a mzIdentifyClass object.
}
\seealso{
The example and demo data of this function can be found 
https://jaspershen.github.io/metID/articles/metID.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
