% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_utils.R
\name{getMS2spectrum}
\alias{getMS2spectrum}
\title{Get MS2 spectra of peaks from databaseClass object}
\usage{
getMS2spectrum(
  lab.id,
  database,
  polarity = c("positive", "negative"),
  ce = "30"
)
}
\arguments{
\item{lab.id}{The lab ID of metabolite.}

\item{database}{Database (databaseClass object).}

\item{polarity}{positive or negative.}

\item{ce}{Collision value.}
}
\value{
A MS2 spectrum (data.frame).
}
\description{
Get MS2 spectra of peaks from databaseClass object.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
