% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{getMS2spectrum2Object}
\alias{getMS2spectrum2Object}
\title{Get spectra of peaks from metIdentifyClass object}
\usage{
getMS2spectrum2Object(object, peak.name)
}
\arguments{
\item{object}{metIdentifyClass.}

\item{peak.name}{Peak name.}
}
\value{
A MS2 spectrum.
}
\description{
Get spectra of peaks from metIdentifyClass object.
\lifecycle{maturing}
}
\seealso{
The example and demo data of this function can be found
https://tidymass.github.io/metid/articles/metid.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
