% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_mzIdentify.R
\name{getParams2}
\alias{getParams2}
\title{Get parameters from a mzIdentifyClass object}
\usage{
getParams2(object)
}
\arguments{
\item{object}{A mzIdentifyClass object.}
}
\value{
A data frame contains all the parameters of this metIdentifiyClass object.
}
\description{
Get parameters from a mzIdentifyClass object.
\lifecycle{deprecated}
}
\seealso{
The example and demo data of this function can be found
https://tidymass.github.io/metid/articles/metid.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
