% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_data.R
\docType{data}
\name{hilic.pos}
\alias{hilic.pos}
\title{HILIC positive mode adduct table}
\format{
A data frame.
}
\usage{
hilic.pos
}
\description{
HILIC positive mode adduct table.
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
\keyword{datasets}
