% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_metabolites_for_multiple_database.R
\name{identify_metabolite_all}
\alias{identify_metabolite_all}
\title{Identify metabolites using multiple databases one time}
\usage{
identify_metabolite_all(ms1.data, ms2.data, parameter.list, path = ".")
}
\arguments{
\item{ms1.data}{The name of ms1 peak table (csv format). Column 1 is "name", column 2 is
"mz" and column 3 is "rt" (second).}

\item{ms2.data}{MS2 data, must be mgf, msp or mzXML format. For example, ms2.data = c("test.mgf", "test2.msp").}

\item{parameter.list}{A list contains paramters for each processing.
The parameter must get using metIdentifyParam or mzIdentifyParam.}

\item{path}{Work directory.}
}
\value{
A list containing mzIdentifyClass object.
}
\description{
Identify metabolites using multiple databases one time.
\lifecycle{maturing}
}
\examples{
\dontrun{
##creat a folder nameed as example
path <- file.path(".", "example")
dir.create(path = path, showWarnings = FALSE)

##get MS1 peak table from metid
ms1_peak <- system.file("ms1_peak", package = "metid")
file.copy(
  from = file.path(ms1_peak, "ms1.peak.table.csv"),
  to = path,
  overwrite = TRUE,
  recursive = TRUE
)

##get MS2 data from metid
ms2_data <- system.file("ms2_data", package = "metid")
file.copy(
  from = file.path(ms2_data, "QC1_MSMS_NCE25.mgf"),
  to = path,
  overwrite = TRUE,
  recursive = TRUE
)

##get databases from metid
database <- system.file("ms2_database", package = "metid")

file.copy(
  from = file.path(
    database,
    c(
      "msDatabase_rplc0.0.2",
      "orbitrapDatabase0.0.1",
      "hmdbMS1Database0.0.1"
    )
  ),
  to = path,
  overwrite = TRUE,
  recursive = TRUE
)
param1 <-
identify_metabolites_params(
  ms1.match.ppm = 15,
  rt.match.tol = 15,
  polarity = "positive",
  ce = "all",
  column = "rp",
  total.score.tol = 0.5,
  candidate.num = 3,
  threads = 3,
  database = "msDatabase_rplc0.0.2"
)

param2 <- identify_metabolites_params(
  ms1.match.ppm = 15,
  rt.match.tol = 15,
  polarity = "positive",
  ce = "all",
  column = "rp",
  total.score.tol = 0.5,
  candidate.num = 3,
  threads = 3,
  database = "orbitrapDatabase0.0.1"
)

param3 <- identify_metabolites_params(
  ms1.match.ppm = 15,
  rt.match.tol = 15,
  polarity = "positive",
  ce = "all",
  column = "rp",
  total.score.tol = 0.5,
  candidate.num = 3,
  threads = 3,
  database = "hmdbMS1Database0.0.1"
)
result <- identify_metabolite_all(
ms1.data = "ms1.peak.table.csv",
ms2.data = "QC1_MSMS_NCE25.mgf",
parameter.list = c(param1, param2, param3),
path = path
)
result[[1]]
result[[2]]
result[[3]]
}
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/multiple_databases.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
