% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{InchiKey}
\alias{InchiKey}
\title{Inchikey database.}
\format{A data frame with 101494 rows and 2 variables:
\describe{
  \item{CID}{PubChem CIDs}
  \item{InchiKey}{Inchikeys}
  ...
}}
\usage{
InchiKey
}
\description{
A dataset containing PubChem CIDs, InchiKey in the PubChem database.
}
\keyword{datasets}
