% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_network}
\alias{kegg_network}
\title{Pairs of kegg network.}
\format{A data frame with 57070 rows and 2 variables:
\describe{
  \item{r1}{KEGG IDs}
  \item{r2}{KEGG IDs, which have a connection with KEGG ID in the first column}
  ...
}}
\usage{
kegg_network
}
\description{
A dataset containing kegg IDs in the KEGG database with all networks.
}
\keyword{datasets}
