% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_deprecated_functions.R
\name{metid}
\alias{metid}
\title{Show the base information of metid pacakge}
\usage{
metid()
}
\value{
A ASCII log of metid
}
\description{
Show the base information of metid pacakge.
}
\examples{
metid()
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
