% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_conflicts.R
\name{metid_conflicts}
\alias{metid_conflicts}
\title{Conflicts between the metid and other packages}
\usage{
metid_conflicts()
}
\value{
metid_conflicts
}
\description{
This function lists all the conflicts between packages in the metid
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
metid_conflicts()
}
