% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_utils.R
\name{metid_packages}
\alias{metid_packages}
\title{List all packages in the metid}
\usage{
metid_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include metid in the list?}
}
\value{
metid_packages
}
\description{
List all packages in the metid
}
\examples{
metid_packages()
}
