% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentify4all.R
\name{metIdentify4all}
\alias{metIdentify4all}
\title{Identify metabolites using multiple databases one time}
\usage{
metIdentify4all(ms1.data, ms2.data, parameter.list, path = ".")
}
\arguments{
\item{ms1.data}{The name of ms1 peak table (csv format). Column 1 is "name", column 2 is
"mz" and column 3 is "rt" (second).}

\item{ms2.data}{MS2 data, must be mgf, msp or mzXML format. For example, ms2.data = c("test.mgf", "test2.msp").}

\item{parameter.list}{A list contains paramters for each processing.
The parameter must get using metIdentifyParam or mzIdentifyParam.}

\item{path}{Work directory.}
}
\value{
A list containing mzIdentifyClass object.
}
\description{
Identify metabolites using multiple databases one time.
\lifecycle{maturing}
}
\seealso{
The example and demo data of this function can be found
https://jaspershen.github.io/metID/articles/metID.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
