% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentifyClass-methods.R
\docType{class}
\name{metIdentifyClass-class}
\alias{metIdentifyClass-class}
\title{An S4 class to represent annotation result.}
\description{
An S4 class to represent annotation result.
}
\section{Slots}{

\describe{
\item{\code{ms1.data}}{MS1 peak table. data.frame.}

\item{\code{ms1.info}}{MS1 information of MS2 data. data.frame.}

\item{\code{ms2.info}}{MS2 information of MS2 data. list.}

\item{\code{identification.result}}{Identification result. list.}

\item{\code{match.result}}{MS1 peak table and MS2 data match result. data.frame.}

\item{\code{adduct.table}}{Adduct table used for annotation. data.frame.}

\item{\code{ms1.ms2.match.mz.tol}}{Parameter for annotation.}

\item{\code{ms1.ms2.match.rt.tol}}{Parameter for annotation.}

\item{\code{ms1.match.ppm}}{Parameter for annotation.}

\item{\code{ms2.match.ppm}}{Parameter for annotation.}

\item{\code{ms2.match.tol}}{Parameter for annotation.}

\item{\code{rt.match.tol}}{Parameter for annotation.}

\item{\code{polarity}}{Parameter for annotation.}

\item{\code{ce}}{Parameter for annotation.}

\item{\code{column}}{Parameter for annotation.}

\item{\code{ms1.match.weight}}{Parameter for annotation.}

\item{\code{rt.match.weight}}{Parameter for annotation.}

\item{\code{ms2.match.weight}}{Parameter for annotation.}

\item{\code{path}}{Parameter for annotation.}

\item{\code{total.score.tol}}{Parameter for annotation.}

\item{\code{candidate.num}}{Parameter for annotation.}

\item{\code{database}}{Parameter for annotation.}

\item{\code{threads}}{Parameter for annotation.}

\item{\code{version}}{Parameter for annotation.}
}}

