% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msDatabase_hilic0.0.2}
\alias{msDatabase_hilic0.0.2}
\title{Michael Snyder lab RPLC database 0.0.2}
\format{
A databaseClass object.
}
\usage{
msDatabase_hilic0.0.2
}
\description{
Michael Snyder lab RPLC database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
\keyword{datasets}
