% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msDatabase_rplc0.0.1}
\alias{msDatabase_rplc0.0.1}
\title{msDatabase_rplc0.0.1}
\format{A databaseClass object.}
\usage{
msDatabase_rplc0.0.1
}
\description{
Michael Snyder lab RPLC database
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
\keyword{datasets}
