% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzIdentify_mass_dataset.R
\name{mzIdentify_mass_dataset2}
\alias{mzIdentify_mass_dataset2}
\title{Identify peaks based on MS1 database}
\usage{
mzIdentify_mass_dataset2(
  object,
  ms1.match.ppm = 25,
  rt.match.tol = 30,
  polarity = c("positive", "negative"),
  column = c("hilic", "rp"),
  candidate.num = 3,
  database,
  threads = 3
)
}
\arguments{
\item{object}{A mass_dataset class object.}

\item{ms1.match.ppm}{Precursor match ppm tolerance.}

\item{rt.match.tol}{RT match tolerance.}

\item{polarity}{The polarity of data, "positive"or "negative".}

\item{column}{"hilic" (HILIC column) or "rp" (reverse phase).}

\item{candidate.num}{The number of candidates.}

\item{database}{MS1 database name or MS1 database.}

\item{threads}{Number of threads}
}
\value{
A mzIdentifyClass or metIdentifyClass object.
}
\description{
Identify peaks based on MS1 database.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
