% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentify4all.R
\name{mzIdentifyParam}
\alias{mzIdentifyParam}
\title{mzIdentify}
\usage{
mzIdentifyParam(
  ms1.match.ppm = 25,
  polarity = c("positive", "negative"),
  column = c("hilic", "rp"),
  candidate.num = 3,
  database,
  threads = 3
)
}
\arguments{
\item{ms1.match.ppm}{Precursor match ppm tolerance.}

\item{polarity}{The polarity of data, "positive"or "negative".}

\item{column}{"hilic" (HILIC column) or "rp" (reverse phase).}

\item{candidate.num}{The number of candidates.}

\item{database}{MS1 database name.}

\item{threads}{Number of threads}
}
\value{
A mzIdentifyClass object.
}
\description{
Generate the mzIdentify parameter list.
}
\seealso{
The example and demo data of this function can be found 
https://jaspershen.github.io/metID/articles/metID.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
