% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_data.R
\docType{data}
\name{orbitrap_database0.0.3}
\alias{orbitrap_database0.0.3}
\title{orbitrap_database0.0.3}
\format{
A databaseClass object.
}
\usage{
orbitrap_database0.0.3
}
\description{
orbitrap_database0.0.3
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
\keyword{datasets}
