% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_mgf.R
\name{read_mgf}
\alias{read_mgf}
\title{read_mgf}
\usage{
read_mgf(file)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be mgf.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MGF data.
\lifecycle{experimental}
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
