% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_deprecated_functions.R
\name{readMGF}
\alias{readMGF}
\title{readMGF}
\usage{
readMGF(file)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be mgf.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MGF data.
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
