% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools4MetaboliteIdentification.R
\name{readMSP}
\alias{readMSP}
\title{readMSP}
\usage{
readMSP(file)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be msp format.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MSP data.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
