% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_msp.R
\name{readMSP_MoNA}
\alias{readMSP_MoNA}
\title{readMSP_MoNA}
\usage{
readMSP_MoNA(file)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be msp. The msp data must from MoNA.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MSP data from MoNA.
}
\details{
---------------------------------------------------------------------------
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
