% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_mzxml.R
\name{readMZXML}
\alias{readMZXML}
\title{readMZXML}
\usage{
readMZXML(file, threads = 3)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be mzXML or mzML.}

\item{threads}{Thread number}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read mzXML data.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
