% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rp.neg}
\alias{rp.neg}
\title{Reverse phase negative mode adduct table}
\format{
A data frame.
}
\usage{
rp.neg
}
\description{
Reverse phase negative mode adduct table.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
\keyword{datasets}
