% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtCorrection.R
\name{rtCor4database}
\alias{rtCor4database}
\title{rtCor4database}
\usage{
rtCor4database(experiment.is.table, database.is.table, database,
  method = c("polyline", "loess"), poly = c(1, 2, 3, 4, 5),
  degree = c(1, 2), path = ".")
}
\arguments{
\item{experiment.is.table}{Experiment internal standard table.}

\item{database.is.table}{Database internal standard table.}

\item{database}{Database, must be datasetClass object.}

\item{method}{polyline or loess.}

\item{poly}{A numeric vector.}

\item{degree}{A numeric vector.}

\item{path}{Work directory.}
}
\value{
A metIdentifyClass object.
}
\description{
Correct RTs of metabolites in database using internal standard list.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
