% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snyder_database_hilic0.0.3}
\alias{snyder_database_hilic0.0.3}
\title{Michael Snyder lab RPLC database}
\format{
A databaseClass object.
}
\usage{
snyder_database_hilic0.0.3
}
\description{
Michael Snyder lab RPLC database
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
\keyword{datasets}
