% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-databaseClass.R
\name{colnames.databaseClass}
\alias{colnames.databaseClass}
\alias{nrow.databaseClass}
\alias{ncol.databaseClass}
\alias{dim.databaseClass}
\title{colnames}
\usage{
\method{colnames}{databaseClass}(x)

\method{nrow}{databaseClass}(x)

\method{ncol}{databaseClass}(x)

\method{dim}{databaseClass}(x)
}
\arguments{
\item{x}{x}
}
\value{
message

message

message

message
}
\description{
colnames

nrow

ncol

dim
}
