% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{trans2newStyle}
\alias{trans2newStyle}
\title{Transform old style identification table to new style}
\usage{
trans2newStyle(identification.table)
}
\arguments{
\item{identification.table}{Identification table from getIdentificationTable or getIdentificationTable2.}
}
\value{
A identification table (data.frame).
}
\description{
Transform old style identification table to new style.
\lifecycle{deprecated}
}
\seealso{
The example and demo data of this function can be found
https://tidymass.github.io/metid/articles/metid.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
