% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_identification_table.R
\name{trans_to_new_style}
\alias{trans_to_new_style}
\title{Transform old style identification table to new style}
\usage{
trans_to_new_style(identification.table)
}
\arguments{
\item{identification.table}{Identification table from get_identification_table.}
}
\value{
A identification table (data.frame).
}
\description{
Transform old style identification table to new style.
\lifecycle{maturing}
}
\examples{
data("annotate_result", package = "metID")
annotation_table <-
get_identification_table(annotate_result, candidate.num = 3, type = "old")
annotation_table <- trans_to_new_style(annotation_table)
}
\seealso{
The example and demo data of this function can be found
https://jaspershen.github.io/metID/articles/metID.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
