% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4metIdentifyClass.R
\name{which_has_identification}
\alias{which_has_identification}
\title{Get the peak names which have identifications}
\usage{
which_has_identification(object)
}
\arguments{
\item{object}{A metIdentifyClass object.}
}
\value{
Peak names with identifications.
}
\description{
Get the peak names which have identifications.
\lifecycle{maturing}
}
\seealso{
The example and demo data of this function can be found
https://tidymass.github.io/metid/articles/metid.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
