% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_for_other_database.R
\name{write_msp_gnps}
\alias{write_msp_gnps}
\title{Export metid database to msp (gnps format)}
\usage{
write_msp_gnps(database, path = ".")
}
\arguments{
\item{database}{metid database.}

\item{path}{Work directory.}
}
\value{
msp format files in local folder.
}
\description{
Export metid database to msp (gnps format).
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
