% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_output_for_other_database.R
\name{write_msp_massbank}
\alias{write_msp_massbank}
\title{Write MassBank Spectra to MSP Format}
\usage{
write_msp_massbank(database, path = ".")
}
\arguments{
\item{database}{An object of class `databaseClass` containing spectral data and metadata.}

\item{path}{A character string specifying the directory where the MSP files will be written. Defaults to the current directory (`"."`).}
}
\value{
This function does not return any value. It writes MSP files for positive and negative mode spectra in the specified directory.
}
\description{
This function writes positive and negative mode MS2 spectra from a given `databaseClass` object to MSP format files. The spectra are written separately for positive and negative ionization modes.
}
\details{
This function extracts the spectra from a `databaseClass` object and writes them into MSP files. The positive mode spectra are written to `spectra_pos.msp` and the negative mode spectra are written to `spectra_neg.msp`. Each entry in the MSP file includes information such as the compound name, precursor type, ion mode, collision energy, and the corresponding MS2 spectrum.
}
\examples{
\dontrun{
# Write MSP data from a databaseClass object
write_msp_massbank(database = my_database, path = "output_directory")
}

}
\seealso{
\code{\link{read_msp}}, \code{\link{write_msp}}
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
