% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseClass-methods.R
\docType{class}
\name{databaseClass-class}
\alias{databaseClass-class}
\title{An S4 class to represent MS1 or MS2 database.}
\description{
An S4 class to represent MS1 or MS2 database.
}
\section{Slots}{

\describe{
\item{\code{database.info}}{Database information.}

\item{\code{spectra.info}}{Metabolites in database.}

\item{\code{spectra.data}}{MS2 spectra data.}
}}

