% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentifyClass-methods.R
\name{filter_adducts}
\alias{filter_adducts}
\title{Filter adducts.}
\usage{
filter_adducts(object, remove_adduct = NULL)
}
\arguments{
\item{object}{metIdentifyClass.}

\item{remove_adduct}{What adduct you want to remove from annotation result. Like '(M-H)-'. All the adduct list can be found here:
data("hilic.pos", package = 'metid'), data("hilic.neg", package = 'metid'),
data("rp.pos", package = 'metid'), data("rp.neg", package = 'metid').}
}
\value{
A MS2 spectrum.
}
\description{
Filter adducts.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
