% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4databaseClass.R
\name{get_ms2_spectrum}
\alias{get_ms2_spectrum}
\title{Get MS2 spectra of peaks from databaseClass object}
\usage{
get_ms2_spectrum(
  lab.id,
  database,
  polarity = c("positive", "negative"),
  ce = "30"
)
}
\arguments{
\item{lab.id}{The lab ID of metabolite.}

\item{database}{Database (databaseClass object).}

\item{polarity}{positive or negative.}

\item{ce}{Collision value.}
}
\value{
A MS2 spectrum (data.frame).

A MS2 spectrum (data.frame).
}
\description{
Get MS2 spectra of peaks from databaseClass object.

Get MS2 spectra of peaks from databaseClass object.
}
\seealso{
The example and demo data of this function can be found
https://tidymass.github.io/metid/articles/metid.html

The example and demo data of this function can be found
https://tidymass.github.io/metid/articles/metid.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}

Xiaotao Shen
\email{shenxt1990@163.com}
}
