% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentifyClass-methods.R
\name{get_ms2_spectrum_from_object}
\alias{get_ms2_spectrum_from_object}
\title{Get spectra of peaks from metIdentifyClass object}
\usage{
get_ms2_spectrum_from_object(object, peak.name)
}
\arguments{
\item{object}{metIdentifyClass.}

\item{peak.name}{Peak name.}
}
\value{
A MS2 spectrum.
}
\description{
Get spectra of peaks from metIdentifyClass object.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
