% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentifyClass-methods.R
\name{get_parameters_metid}
\alias{get_parameters_metid}
\title{Get parameters from a metIdentifyClass object}
\usage{
get_parameters_metid(object)
}
\arguments{
\item{object}{A metIdentifyClass object.}
}
\value{
A data frame contains all the parameters of this metIdentifiyClass object.
}
\description{
Get parameters from a metIdentifyClass object.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
