% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{getIdenInfo}
\alias{getIdenInfo}
\title{Get identification information from a metIdentifyClass object}
\usage{
getIdenInfo(object, which.peak, database)
}
\arguments{
\item{object}{A metIdentifyClass object.}

\item{which.peak}{A peak name or "all". "all" means all peaks with identifications will be output.}

\item{database}{Database used.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification information from a metIdentifyClass object.
\lifecycle{deprecated}
}
\examples{
data("annotate_result", package = "metID")
data("msDatabase_hilic0.0.2", package = "metID")
peak_name <- which_has_identification(annotate_result)
getIdenInfo(annotate_result, peak_name[1,1], database = msDatabase_rplc0.0.2)
}
\seealso{
The example and demo data of this function can be found
https://jaspershen.github.io/metID/articles/metID.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
