% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4metIdentifyClass.R
\name{getIdenInfo}
\alias{getIdenInfo}
\title{getIdenInfo}
\usage{
getIdenInfo(object, which.peak, database)
}
\arguments{
\item{object}{A metIdentifyClass object.}

\item{which.peak}{A peak name or "all". "all" means all peaks with identifications will be output.}

\item{database}{Database used.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification information from a metIdentifyClass object.
}
\seealso{
The example and demo data of this function can be found 
https://jaspershen.github.io/metID/articles/metID.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
