% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{getIdentificationTable}
\alias{getIdentificationTable}
\title{Get identification table from a metIdentifyClass object}
\usage{
getIdentificationTable(..., candidate.num = 3, type = c("old", "new"))
}
\arguments{
\item{...}{One or multiple metIdentifyClass objects.}

\item{candidate.num}{The number of candidates.}

\item{type}{The type of identification table.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification table from a metIdentifyClass object.
\lifecycle{deprecated}
}
\examples{
data("annotate_result", package = "metID")
annotation_table <- 
getIdentificationTable(annotate_result, candidate.num = 3, type = "old")
annotation_table
annotation_table <- 
getIdentificationTable(annotate_result, candidate.num = 3, type = "new")
annotation_table
}
\seealso{
The example and demo data of this function can be found
https://jaspershen.github.io/metID/articles/metID.html
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
