% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_identification_table.R
\name{getIdentificationTable2}
\alias{getIdentificationTable2}
\title{Get identification table from a mzIdentifyClass object}
\usage{
getIdentificationTable2(
  object,
  candidate.num = 3,
  type = c("old", "new"),
  silence.deprecated = FALSE
)
}
\arguments{
\item{object}{mzIdentifyClass object.}

\item{candidate.num}{The number of candidates.}

\item{type}{The type of identification table.}

\item{silence.deprecated}{Silenc the deprecated information or not.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification table from a mzIdentifyClass object.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
