% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hilic.neg}
\alias{hilic.neg}
\title{HILIC negative mode adduct table}
\format{
A data frame.
}
\usage{
hilic.neg
}
\description{
HILIC negative mode adduct table.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
\keyword{datasets}
