% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_metabolites_for_multiple_database.R
\name{identify_metabolite_all}
\alias{identify_metabolite_all}
\title{Identify metabolites using multiple databases one time}
\usage{
identify_metabolite_all(ms1.data, ms2.data, parameter.list, path = ".")
}
\arguments{
\item{ms1.data}{The name of ms1 peak table (csv format). Column 1 is "name", column 2 is
"mz" and column 3 is "rt" (second).}

\item{ms2.data}{MS2 data, must be mgf, msp or mzXML format. For example, ms2.data = c("test.mgf", "test2.msp").}

\item{parameter.list}{A list contains paramters for each processing.
The parameter must get using metIdentifyParam or mzIdentifyParam.}

\item{path}{Work directory.}
}
\value{
A list containing mzIdentifyClass object.
}
\description{
Identify metabolites using multiple databases one time.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/multiple_databases.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
