% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools4MetaboliteIdentification.R
\name{identify_peak}
\alias{identify_peak}
\title{Identify metabolites based on MS1 or MS/MS database}
\usage{
identify_peak(
  idx,
  ms1.info,
  ms2.info,
  spectra.info,
  spectra.data,
  ppm.ms1match = 25,
  ppm.ms2match = 30,
  mz.ppm.thr = 400,
  ms2.match.tol = 0.5,
  rt.match.tol = 30,
  ms1.match.weight = 0.25,
  rt.match.weight = 0.25,
  ms2.match.weight = 0.5,
  total.score.tol = 0.5,
  adduct.table,
  candidate.num = 3,
  fraction.weight = 0.3,
  dp.forward.weight = 0.6,
  dp.reverse.weight = 0.1,
  remove_fragment_intensity_cutoff = 0,
  ...
)
}
\arguments{
\item{idx}{idx}

\item{ms1.info}{ms1.info}

\item{ms2.info}{ms2.info}

\item{spectra.info}{spectra.info}

\item{spectra.data}{spectra.data}

\item{ppm.ms1match}{ppm.ms1match}

\item{ppm.ms2match}{ppm.ms2match}

\item{mz.ppm.thr}{Accurate mass tolerance for m/z error calculation.}

\item{ms2.match.tol}{MS2 match (MS2 similarity) tolerance.}

\item{rt.match.tol}{The weight for matched fragments.}

\item{ms1.match.weight}{Forward dot product weight.}

\item{rt.match.weight}{Reverse dot product weight.}

\item{ms2.match.weight}{RT match tolerance.}

\item{total.score.tol}{The polarity of data, "positive"or "negative".}

\item{adduct.table}{Collision energy. Please confirm the CE values in your database. Default is "all".}

\item{candidate.num}{"hilic" (HILIC column) or "rp" (reverse phase).}

\item{fraction.weight}{The weight of MS1 match for total score calculation.}

\item{dp.forward.weight}{The weight of RT match for total score calculation.}

\item{dp.reverse.weight}{The weight of MS2 match for total score calculation.}

\item{remove_fragment_intensity_cutoff}{remove_fragment_intensity_cutoff}

\item{...}{other parameters}
}
\value{
A metIdentifyClass object.
}
\description{
Identify metabolites based on MS1 or MS/MS database.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
