% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53_ms2_spectra_matching_scores.R
\name{match_ms2_fragments}
\alias{match_ms2_fragments}
\title{Match MS2 Fragments Between Experimental and Library Spectra}
\usage{
match_ms2_fragments(
  experimental.spectrum,
  library.spectrum,
  ms2.match.ppm = 30,
  mz.ppm.thr = 400,
  direction = c("reverse", "forward"),
  remove.noise = TRUE
)
}
\arguments{
\item{experimental.spectrum}{A data frame representing the experimental spectrum, with columns `mz` for mass-to-charge ratios and `intensity` for corresponding intensities.}

\item{library.spectrum}{A data frame representing the library spectrum, with columns `mz` for mass-to-charge ratios and `intensity` for corresponding intensities.}

\item{ms2.match.ppm}{Numeric. The mass tolerance in parts per million (ppm) for matching MS2 fragments. Default is 30.}

\item{mz.ppm.thr}{Numeric. A threshold for mass-to-charge ratio (m/z) to use in ppm calculation. Default is 400.}
}
\value{
A data frame with matched MS2 fragments, containing the following columns:
\describe{
  \item{Lib.index}{The index of the fragment in the library spectrum.}
  \item{Exp.index}{The index of the matched fragment in the experimental spectrum.}
  \item{Lib.mz}{The m/z value of the fragment in the library spectrum.}
  \item{Lib.intensity}{The intensity of the fragment in the library spectrum.}
  \item{Exp.mz}{The m/z value of the matched fragment in the experimental spectrum.}
  \item{Exp.intensity}{The intensity of the matched fragment in the experimental spectrum.}
}
}
\description{
This function matches MS2 fragments between an experimental spectrum and a library spectrum based on m/z values, using a specified mass tolerance (ppm). It returns a matrix of matched fragments, including their m/z values and intensities.
}
\details{
The function first removes noisy fragments from both the experimental and library spectra based on the provided mass tolerance (ppm). It then performs matching by comparing the m/z values of the fragments. If multiple experimental fragments match a library fragment, the one with the highest intensity is selected. Any unmatched fragments are added with NA values for the respective library or experimental fragment.
}
\author{
Xiaotao Shen \email{xiaotao.shen@outlook.com}
}
