% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID.R
\name{metID}
\alias{metID}
\title{Show the base information of metID pacakge}
\usage{
metID()
}
\value{
A ASCII log of metID
}
\description{
Show the base information of metID pacakge.
\lifecycle{maturing}
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
