% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{MetID}
\alias{MetID}
\alias{MetID-package}
\title{MetID: A package for Network-based prioritization of putative metabolite IDs.}
\description{
The foo package provides one important functions:
get_scores_for_LC_MS
}
\section{Foo functions}{

get_scores_for_LC_MS: Get scores for metabolite putative IDs by LC-MS.
}

