% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metid.R
\name{metid}
\alias{metid}
\title{Show the base information of metid pacakge}
\usage{
metid()
}
\value{
A ASCII log of metid
}
\description{
Show the base information of metid pacakge.
}
\examples{
metid()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
