% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metID_logo.R
\name{metID_logo}
\alias{metID_logo}
\title{Show the base information of metID pacakge}
\usage{
metID_logo()
}
\value{
A ASCII log of metID
}
\description{
Show the base information of metID pacakge.
\lifecycle{maturing}
}
\examples{
metID_logo()
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
