% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_metid_logo.R
\name{metid_logo}
\alias{metid_logo}
\title{Show the base information of metid pacakge}
\usage{
metid_logo()
}
\value{
A ASCII log of metid
}
\description{
Show the base information of metid pacakge.
}
\examples{
metid_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
