% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_plot_mass_dataset.R
\name{ms2_plot_mass_dataset}
\alias{ms2_plot_mass_dataset}
\title{Identify metabolites based on MS1 or MS/MS database for single peak.}
\usage{
ms2_plot_mass_dataset(
  object,
  variable_id,
  variable_index,
  polarity = c("positive", "negative"),
  ms1.match.ppm = 25,
  ms2.match.ppm = 30,
  mz.ppm.thr = 400,
  database,
  interactive_plot = FALSE
)
}
\arguments{
\item{object}{A mass_dataset class obejct.}

\item{variable_id}{variable_id}

\item{variable_index}{variable_index}

\item{polarity}{The polarity of data, "positive"or "negative".}

\item{ms1.match.ppm}{Precursor match ppm tolerance.}

\item{ms2.match.ppm}{Fragment ion match ppm tolerance.}

\item{mz.ppm.thr}{Accurate mass tolerance for m/z error calculation.}

\item{database}{MS2 database name or MS database.}

\item{interactive_plot}{interactive_plot}
}
\value{
A list of ggplot2 object.
}
\description{
Identify metabolites based on MS1 or MS/MS database.
}
\seealso{
The example and demo data of this function can be found
\url{https://tidymass.github.io/metid/articles/metid.html}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
