% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_mgf.R
\name{read_mgf_experiment}
\alias{read_mgf_experiment}
\title{read_mgf_experiment}
\usage{
read_mgf_experiment(file, threads = 3)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be mgf.}

\item{threads}{threads number.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MGF data from experiment.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
