% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_msp.R
\name{read_msp}
\alias{read_msp}
\title{read_msp}
\usage{
read_msp(file, threads = 3)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be msp format.}

\item{threads}{threads number.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MSP data.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
