% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_msp.R
\name{read_msp_database}
\alias{read_msp_database}
\title{read_msp_database}
\usage{
read_msp_database(file, threads = 4)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be msp format.}

\item{threads}{threads}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MSP data for database.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
