% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_msp.R
\name{read_msp_gnps}
\alias{read_msp_gnps}
\title{read_msp_gnps}
\usage{
read_msp_gnps(file, threads = 5)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be msp. The msp data must from MoNA.}

\item{threads}{The number of threads}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MSP data from MoNA.
}
\details{
---------------------------------------------------------------------------
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
