% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools4MetaboliteIdentification.R
\name{readMGF}
\alias{readMGF}
\title{readMGF}
\usage{
readMGF(file)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be mgf.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MGF data.
\lifecycle{experimental}
}
\examples{
ms2_data <- system.file("ms2_data", package = "metID")
dir(ms2_data)
data <- readMGF(dir(ms2_data, full.names = TRUE))
data[[1]]
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
