% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_msp.R
\name{readMSP}
\alias{readMSP}
\title{readMSP}
\usage{
readMSP(file, silence.deprecated = TRUE)
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be msp format.}

\item{silence.deprecated}{Silenc the deprecated information or not.}
}
\value{
Return ms2 data. This is a list.
}
\description{
Read MSP data.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
