% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_utils.R
\name{summary_annotation_table}
\alias{summary_annotation_table}
\title{Summary the annotation  table}
\usage{
summary_annotation_table(object, level = c(1, 2, 3, 4))
}
\arguments{
\item{object}{mass_dataset class or data.frame.}

\item{level}{what levels you want to use.}
}
\value{
A ggplot2 object
}
\description{
Summary the annotation  table
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
