% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_for_other_database.R
\name{write_mgf_massbank}
\alias{write_mgf_massbank}
\title{Export metid database to mgf (MassBank format)}
\usage{
write_mgf_massbank(database, path = ".")
}
\arguments{
\item{database}{metid database.}

\item{path}{Work directory.}
}
\value{
mgf format files in local folder.
}
\description{
Export metid database to mgf (MassBank format).
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
