% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_for_other_database.R
\name{write_msp_mona}
\alias{write_msp_mona}
\title{Export metID database to msp (mona format)}
\usage{
write_msp_mona(databasae, path = ".")
}
\arguments{
\item{databasae}{metID database.}

\item{path}{Work directory.}
}
\value{
msp format files in local folder.
}
\description{
Export metID database to msp (mona format).
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
